rename curl -> fech
