--- applications/system/curl/application.fam
@@ -1,9 +1,0 @@
-App(
-    appid="curl",
-    name="curl",
-    targets=["f20", "f21"],
-    apptype=FlipperAppType.CLICMD,
-    entry_point="curl_command",
-    stack_size=1024 * 2,
-    sources=["curl.c", "helpers/curl_client.c"],
-)
