--- applications/system/curl/curl.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <cli/cli_command.h>
-
-void curl_command(PipeSide* pipe, FuriString* args, void* context);
