+++ applications/system/fetch/fetch.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <cli/cli_command.h>
+
+void fetch_command(PipeSide* pipe, FuriString* args, void* context);
